+++ tests/i18n/test_funcs.py
@@ -3,8 +3,11 @@ from cjwparse.i18n import _trans_cjwparse
 
 
 def test_trans_cjwparse():
-    assert _trans_cjwparse(
-        "errors.allNull",
-        "The column “{column}” must contain non-null values.",
-        {"column": "A"},
-    ) == I18nMessage("errors.allNull", {"column": "A"}, "cjwparse")
+    assert (
+        _trans_cjwparse(
+            "errors.allNull",
+            "The column “{column}” must contain non-null values.",
+            {"column": "A"},
+        )
+        == I18nMessage("errors.allNull", {"column": "A"}, "cjwparse")
+    )