+++ tests/util.py
@@ -1,28 +1,6 @@
-import tempfile
-import unittest
-from contextlib import contextmanager
-from pathlib import Path
-from typing import Any, ContextManager, Dict, List, Optional, Union
+from typing import Any, Dict, Union
 
 import pyarrow
-from numpy.testing import assert_equal  # [None, "x"] == [None, "x"]
-
-
-@contextmanager
-def arrow_file(
-    table: Union[Dict[str, List[Any]], pyarrow.Table], dir: Optional[Path] = None
-) -> ContextManager[Path]:
-    """
-    Yield a path with `table` written to an Arrow file.
-    """
-    if isinstance(table, dict):
-        table = pyarrow.table(table)
-
-    with tempfile.NamedTemporaryFile(dir=dir) as tf:
-        writer = pyarrow.RecordBatchFileWriter(tf.name, table.schema)
-        writer.write_table(table)
-        writer.close()
-        yield Path(tf.name)
 
 
 def assert_arrow_table_equals(
@@ -30,17 +8,10 @@ def assert_arrow_table_equals(
 ) -> None:
     if isinstance(result2, dict):
         result2 = pyarrow.table(result2)
-    assertEqual = unittest.TestCase().assertEqual
-    assertEqual(result1.shape, result2.shape)
-    assertEqual(result1.column_names, result2.column_names)
+    assert result1.shape == result2.shape
+    assert result1.column_names == result2.column_names
     for colname, actual_col, expected_col in zip(
         result1.column_names, result1.columns, result2.columns
     ):
-        assertEqual(
-            actual_col.type, expected_col.type, msg=f"Column {colname} has wrong type"
-        )
-        assert_equal(
-            actual_col.to_pylist(),
-            expected_col.to_pylist(),
-            err_msg=f"Column {colname} has wrong values",
-        )
+        assert actual_col.type == expected_col.type
+        assert actual_col.to_pylist() == expected_col.to_pylist()